requirements.txt
streamlit
faiss-cpu
sentence-transformers
rank-bm25
